fix: revert back to showadd prop
