+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/blocks/Image/Image.spec.tsx
@@ -69,7 +69,7 @@ describe('Image', () => {
     expect(dispatch).toHaveBeenCalledWith({
       type: 'SetDrawerPropsAction',
       title: 'Image',
-      children: <ImageOptions noSource />
+      children: <ImageOptions />
     })
   })
 })
